move ensureFileOwner back to the ssm impl file and tweak lease timeout (#8)
